fix(Makefile): All build to fail on test failure
The `test/run_tests.sh` would not allow the build to fail for failed
unit test cases. This was problematic given that the CI pipeline catches
bad unit tests through failed unit test executables.

From now on, we will just name the unit test we want to execute in the
Makefile directly instead of using an external script.

Signed-off-by: Jason Pena <jasonpena@awkless.com>